requirements.txt
streamlit
langchain
openai
chromadb
tiktoken
pypdf
pinecone-client
